@@ -1,7 +1,7 @@
 module TradingAppShared
 {
     requires java.sql;
-    exports Models;
+    exports Shared_Models;
     exports Enums;
     exports Messages;
 
